Refactoring: split views.py, update urls.py #34
