+++ backend/urls.py
@@ -1,35 +1,65 @@
 from django.urls import path
 
-from . import views
+from .views import (
+    check_in,
+    confirm,
+    attendance_list,
+    classes,
+    today_classes_list,
+    today_class_occurrences,
+    class_occurrences,
+    delete_occurrence,
+    edit_occurrence,
+    edit_class,
+    delete_class,
+    students_view,
+    list_students,
+    create_student,
+    edit_student,
+    delete_student,
+    get_attended_students,
+    prices,
+    edit_price,
+    payments,
+    delete_payment,
+    payment_summary,
+    schedules,
+    delete_schedule,
+    available_time_slots,
+    get_user,
+    schools,
+    school_detail,
+    edit_school,
+    delete_school,
+)
 
 urlpatterns = [
-    path("check_in/", views.check_in, name="check_in"),
-    path("confirm/", views.confirm, name="confirm"),
-    path("attendances/", views.attendance_list, name="attendances"),
-    path("classes/", views.classes, name="classes"),
-    path("today_classes_list/", views.today_classes_list, name="today_classes_list"),
-    path("today_class_occurrences/", views.today_class_occurrences, name="today_class_occurrences"),
-    path("class_occurrences/", views.class_occurrences, name="class_occurrences"),
-    path("class_occurrences/<int:occurrence_id>/delete/", views.delete_occurrence, name="delete_occurrence"),
-    path("class_occurrences/<int:occurrence_id>/edit/", views.edit_occurrence, name="edit_occurrence"),
-    path("classes/<int:class_id>/edit/", views.edit_class, name="edit_class"),
-    path("classes/<int:class_id>/delete/", views.delete_class, name="delete_class"),
-    path("students/", views.students_view, name="students"),
-    path("students/<int:student_id>/edit/", views.edit_student, name="edit_student"),
-    path("students/<int:student_id>/delete/", views.delete_student, name="delete_student"),
-    path("attended_students/", views.get_attended_students, name="attended_students"),
-    path("prices/", views.prices, name="prices"),
-    path("prices/<int:price_id>/", views.edit_price, name="edit_price"),
-    path("payments/", views.payments, name="payments"),
-    path("payments/<int:payment_id>/delete/", views.delete_payment, name="delete_payment"),
-    path("payment_summary/", views.payment_summary, name="payment_summary"),
-    path("schedules/", views.schedules, name="schedules"),
-    path("schedules/<int:schedule_id>/delete/", views.delete_schedule, name="delete_schedule"),
-    path("available_time_slots/", views.available_time_slots, name="available_time_slots"),
-    path("available_occurrence_time/", views.available_occurrence_time, name="available_occurrence_time"),
-    path("me/", views.get_user, name="get_user"),
-    path("schools/", views.schools, name="schools"),
-    path("schools/<int:school_id>/", views.school_detail, name="school_detail"),
-    path("schools/<int:school_id>/edit/", views.edit_school, name="edit_school"),
-    path("schools/<int:school_id>/delete/", views.delete_school, name="delete_school"),
+    path("check_in/", check_in, name="check_in"),
+    path("confirm/", confirm, name="confirm"),
+    path("attendances/", attendance_list, name="attendances"),
+    path("classes/", classes, name="classes"),
+    path("today_classes_list/", today_classes_list, name="today_classes_list"),
+    path("today_class_occurrences/", today_class_occurrences, name="today_class_occurrences"),
+    path("class_occurrences/", class_occurrences, name="class_occurrences"),
+    path("class_occurrences/<int:occurrence_id>/delete/", delete_occurrence, name="delete_occurrence"),
+    path("class_occurrences/<int:occurrence_id>/edit/", edit_occurrence, name="edit_occurrence"),
+    path("classes/<int:class_id>/edit/", edit_class, name="edit_class"),
+    path("classes/<int:class_id>/delete/", delete_class, name="delete_class"),
+    path("students/", students_view, name="students"),
+    path("students/<int:student_id>/edit/", edit_student, name="edit_student"),
+    path("students/<int:student_id>/delete/", delete_student, name="delete_student"),
+    path("attended_students/", get_attended_students, name="attended_students"),
+    path("prices/", prices, name="prices"),
+    path("prices/<int:price_id>/", edit_price, name="edit_price"),
+    path("payments/", payments, name="payments"),
+    path("payments/<int:payment_id>/delete/", delete_payment, name="delete_payment"),
+    path("payment_summary/", payment_summary, name="payment_summary"),
+    path("schedules/", schedules, name="schedules"),
+    path("schedules/<int:schedule_id>/delete/", delete_schedule, name="delete_schedule"),
+    path("available_time_slots/", available_time_slots, name="available_time_slots"),
+    path("me/", get_user, name="get_user"),
+    path("schools/", schools, name="schools"),
+    path("schools/<int:school_id>/", school_detail, name="school_detail"),
+    path("schools/<int:school_id>/edit/", edit_school, name="edit_school"),
+    path("schools/<int:school_id>/delete/", delete_school, name="delete_school"),
 ]
